pyproject.toml
[tool.poetry]
name = "tokenapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymongo="^4.6.1"
Requests="^2.31.0"
web3="^6.12.0"
python-dotenv = "^1.0.0"
pytz = "^2023.3.post1"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
